Settings: fixup! Wi-Fi timeout feature
Issue: calyxos#825

+++ src/com/android/settings/wifi/WifiTimeoutPreferenceController.java
@@ -72,8 +72,8 @@ public class WifiTimeoutPreferenceController extends BasePreferenceController im
     @Override
     public boolean onPreferenceChange(Preference preference, Object newValue) {
         try {
-            int value = Integer.parseInt((String) newValue);
-            Settings.Global.putInt(mContext.getContentResolver(), Settings.Global.WIFI_OFF_TIMEOUT, value);
+            long value = Long.parseLong((String) newValue);
+            Settings.Global.putLong(mContext.getContentResolver(), Settings.Global.WIFI_OFF_TIMEOUT, value);
             updateTimeoutPreferenceDescription((ListPreference) preference, value);
         } catch (NumberFormatException e) {
             Log.e(TAG, "could not persist wifi timeout setting", e);
